+++ events_test.go
@@ -12,13 +12,13 @@ func TestEvents_String(t *testing.T) {
 		out   string
 	}{
 		// RunEvent
-		{RunEvent{User: "ejholmes", App: "acme-inc", Command: []string{"bash"}}, "ejholmes started `bash` (detached) on acme-inc"},
+		{RunEvent{User: "ejholmes", App: "acme-inc", Command: []string{"bash"}}, "ejholmes started running `bash` (detached) on acme-inc"},
 		{RunEvent{User: "ejholmes", App: "acme-inc", Command: []string{"bash"}, Finished: true}, "ejholmes ran `bash` (detached) on acme-inc"},
-		{RunEvent{User: "ejholmes", App: "acme-inc", Attached: true, Command: []string{"bash"}}, "ejholmes started `bash` (attached) on acme-inc"},
-		{RunEvent{User: "ejholmes", App: "acme-inc", URL: "https://console.aws.amazon.com/cloudwatch/home?region=us-east-1#logEvent:group=runs;stream=dac6eaff-6e0b-4708-9277-9f38aea2f528", Attached: true, Command: []string{"bash"}}, "ejholmes started `bash` (attached) on acme-inc (<https://console.aws.amazon.com/cloudwatch/home?region=us-east-1#logEvent:group=runs;stream=dac6eaff-6e0b-4708-9277-9f38aea2f528|logs>)"},
-		{RunEvent{User: "ejholmes", App: "acme-inc", Command: []string{"bash"}, Message: "commit message"}, "ejholmes started `bash` (detached) on acme-inc: 'commit message'"},
-		{RunEvent{User: "ejholmes", App: "acme-inc", Attached: true, Command: []string{"bash"}, Message: "commit message"}, "ejholmes started `bash` (attached) on acme-inc: 'commit message'"},
-		{RunEvent{User: "ejholmes", App: "acme-inc", URL: "https://console.aws.amazon.com/cloudwatch/home?region=us-east-1#logEvent:group=runs;stream=dac6eaff-6e0b-4708-9277-9f38aea2f528", Attached: true, Command: []string{"bash"}, Message: "commit message"}, "ejholmes started `bash` (attached) on acme-inc (<https://console.aws.amazon.com/cloudwatch/home?region=us-east-1#logEvent:group=runs;stream=dac6eaff-6e0b-4708-9277-9f38aea2f528|logs>): 'commit message'"},
+		{RunEvent{User: "ejholmes", App: "acme-inc", Attached: true, Command: []string{"bash"}}, "ejholmes started running `bash` (attached) on acme-inc"},
+		{RunEvent{User: "ejholmes", App: "acme-inc", URL: "https://console.aws.amazon.com/cloudwatch/home?region=us-east-1#logEvent:group=runs;stream=dac6eaff-6e0b-4708-9277-9f38aea2f528", Attached: true, Command: []string{"bash"}}, "ejholmes started running `bash` (attached) on acme-inc (<https://console.aws.amazon.com/cloudwatch/home?region=us-east-1#logEvent:group=runs;stream=dac6eaff-6e0b-4708-9277-9f38aea2f528|logs>)"},
+		{RunEvent{User: "ejholmes", App: "acme-inc", Command: []string{"bash"}, Message: "commit message"}, "ejholmes started running `bash` (detached) on acme-inc: 'commit message'"},
+		{RunEvent{User: "ejholmes", App: "acme-inc", Attached: true, Command: []string{"bash"}, Message: "commit message"}, "ejholmes started running `bash` (attached) on acme-inc: 'commit message'"},
+		{RunEvent{User: "ejholmes", App: "acme-inc", URL: "https://console.aws.amazon.com/cloudwatch/home?region=us-east-1#logEvent:group=runs;stream=dac6eaff-6e0b-4708-9277-9f38aea2f528", Attached: true, Command: []string{"bash"}, Message: "commit message"}, "ejholmes started running `bash` (attached) on acme-inc (<https://console.aws.amazon.com/cloudwatch/home?region=us-east-1#logEvent:group=runs;stream=dac6eaff-6e0b-4708-9277-9f38aea2f528|logs>): 'commit message'"},
 
 		// RestartEvent
 		{RestartEvent{User: "ejholmes", App: "acme-inc"}, "ejholmes restarted acme-inc"},